Move /access endpoint & Add invalid userdata testcase
