+++ src/app/routes/routes.ts
@@ -1,5 +1,4 @@
 import { Router } from 'express';
-import speakeasy from 'speakeasy';
 import createHttpError from 'http-errors';
 
 import AuthRoutes from './auth';
@@ -34,7 +33,7 @@ export default (router: Router, service: any, App: any): Router => {
   service.Analytics = AnalyticsService;
   service.ReCaptcha = ReCaptchaV3;
 
-  AuthRoutes(router, service);
+  AuthRoutes(router, service, App.config);
   ActivationRoutes(router, service);
   StorageRoutes(router, service, App);
   BridgeRoutes(router, service);
@@ -54,112 +53,6 @@ export default (router: Router, service: any, App: any): Router => {
   NewsletterRoutes(router, service);
   UserRoutes(router, service, App);
   AnalyticsRoutes(router);
-
-  router.post('/access', async (req, res) => {
-    const MAX_LOGIN_FAIL_ATTEMPTS = 10;
-
-    const userData: any = await service.User.FindUserByEmail(req.body.email);
-
-    if (!userData) {
-      throw createHttpError(401, 'Wrong email/password');
-    }
-
-    const loginAttemptsLimitReached = userData.errorLoginCount >= MAX_LOGIN_FAIL_ATTEMPTS;
-
-    if (loginAttemptsLimitReached) {
-      throw createHttpError(401, 'Your account has been blocked for security reasons. Please reach out to us');
-    }
-
-    const hashedPass = App.services.Crypt.decryptText(req.body.password);
-
-    if (hashedPass !== userData.password.toString()) {
-      service.User.LoginFailed(req.body.email, true);
-      throw createHttpError(401, 'Wrong email/password');
-    }
-
-    const twoFactorEnabled = userData.secret_2FA && userData.secret_2FA.length > 0;
-    if (twoFactorEnabled) {
-      const tfaResult = speakeasy.totp.verifyDelta({
-        secret: userData.secret_2FA,
-        token: req.body.tfa,
-        encoding: 'base32',
-        window: 2,
-      });
-
-      if (!tfaResult) {
-        throw createHttpError(401, 'Wrong 2-factor auth code');
-      }
-    }
-
-    const internxtClient = req.headers['internxt-client'];
-    const token = Sign(userData.email, App.config.get('secrets').JWT, internxtClient === 'drive-web');
-    const newToken = Sign({ payload: {
-      uuid: userData.uuid,
-      email: userData.email,
-      name: userData.name,
-      lastname: userData.lastname,
-      username: userData.username,
-      sharedWorkspace: true,
-      networkCredentials: {
-        user: userData.bridgeUser,
-        pass: userData.userId
-      }
-    } }, App.config.get('secrets').JWT);
-
-    service.User.LoginFailed(req.body.email, false);
-    service.User.UpdateAccountActivity(req.body.email);
-
-    const userBucket = await service.User.GetUserBucket(userData);
-    const keyExists = await service.KeyServer.keysExists(userData);
-
-    if (!keyExists && req.body.publicKey) {
-      await service.KeyServer.addKeysLogin(userData, req.body.publicKey, req.body.privateKey, req.body.revocateKey);
-    }
-
-    const keys = await service.KeyServer.getKeys(userData);
-    const hasTeams = !!(await service.Team.getTeamByMember(req.body.email));
-    const appSumoDetails = await service.AppSumo.GetDetails(userData.id).catch(() => null);
-
-    const user = {
-      email: req.body.email,
-      userId: userData.userId,
-      mnemonic: userData.mnemonic,
-      root_folder_id: userData.root_folder_id,
-      name: userData.name,
-      lastname: userData.lastname,
-      uuid: userData.uuid,
-      credit: userData.credit,
-      createdAt: userData.createdAt,
-      privateKey: keys ? keys.private_key : null,
-      publicKey: keys ? keys.public_key : null,
-      revocateKey: keys ? keys.revocation_key : null,
-      bucket: userBucket,
-      registerCompleted: userData.registerCompleted,
-      teams: hasTeams,
-      username: userData.username,
-      bridgeUser: userData.bridgeUser,
-      sharedWorkspace: userData.sharedWorkspace,
-      appSumoDetails: appSumoDetails || null,
-      hasReferralsProgram: await service.UsersReferrals.hasReferralsProgram(
-        userData.id,
-        userData.email,
-        userData.bridgeUser,
-        userData.userId,
-      ),
-      backupsBucket: userData.backupsBucket,
-    };
-
-    const userTeam = null;
-    // TODO: Not working. Team members can not use team workspace due to this
-    // if (userTeam) {
-    //   const tokenTeam = Sign(userTeam.bridge_user, App.config.get('secrets').JWT, internxtClient === 'drive-web');
-
-    //   return res.status(200).json({
-    //     user, token, userTeam, tokenTeam
-    //   });
-    // }
-    return res.status(200).json({ user, token, userTeam, newToken });
-  });
 
   router.get('/user/refresh', passportAuth, async (req, res) => {
     const { publicKey, privateKey, revocateKey } = req.body;
